FEAT: Implement profile page and related UI components, update navigation for improved user experience

+++ lib/core/providers/wishlist_provider.dart
@@ -1,14 +1,22 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:movie_pedia/core/models/movie_detail_model.dart';
+import 'package:movie_pedia/core/providers/firebase_auth_provider.dart';
 
 final wishlistProvider =
     StateNotifierProvider<WishlistNotifier, List<WishlistMovie>>((ref) {
-  return WishlistNotifier();
+  final user = ref.watch(authStateNotifierProvider).value;
+  return WishlistNotifier(userId: user?.uid);
+});
+
+final wishlistCountProvider = Provider<AsyncValue<int>>((ref) {
+  final wishlistState = ref.watch(wishlistProvider);
+  return AsyncValue.data(wishlistState.length);
 });
 
 class WishlistMovie {
   final String id;
+  final String userId;
   final String title;
   final String posterPath;
   final double voteAverage;
@@ -17,6 +25,7 @@ class WishlistMovie {
 
   WishlistMovie({
     required this.id,
+    required this.userId,
     required this.title,
     required this.posterPath,
     required this.voteAverage,
@@ -26,6 +35,7 @@ class WishlistMovie {
 
   Map<String, dynamic> toMap() {
     return {
+      'userId': userId,
       'title': title,
       'posterPath': posterPath,
       'voteAverage': voteAverage,
@@ -37,6 +47,7 @@ class WishlistMovie {
   static WishlistMovie fromMap(String id, Map<String, dynamic> data) {
     return WishlistMovie(
       id: id,
+      userId: data['userId'] ?? '',
       title: data['title'],
       posterPath: data['posterPath'],
       voteAverage: (data['voteAverage']).toDouble(),
@@ -50,36 +61,82 @@ class WishlistMovie {
 
 class WishlistNotifier extends StateNotifier<List<WishlistMovie>> {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
+  final String? userId; // Add userId field
 
-  WishlistNotifier() : super([]) {
-    loadWishlist();
+  WishlistNotifier({this.userId}) : super([]) {
+    if (userId != null) {
+      loadWishlist();
+    }
   }
 
   Future<void> loadWishlist() async {
-    final snapshot = await _firestore.collection('wishlist').get();
+    if (userId == null) return;
+
+    final snapshot = await _firestore
+        .collection('wishlist')
+        .where('userId', isEqualTo: userId)
+        .get();
+
     state = snapshot.docs
         .map((doc) => WishlistMovie.fromMap(doc.id, doc.data()))
         .toList();
   }
 
+  Future<int?> getWishlistCount() async {
+    if (userId == null) return 0;
+
+    final snapshot = await _firestore
+        .collection('wishlist')
+        .where('userId', isEqualTo: userId)
+        .count()
+        .get();
+
+    return snapshot.count;
+  }
+
   Future<void> addToWishlist(MovieDetailModel movieDetail) async {
+    if (userId == null) return;
+
     final exists = state.any((m) => m.title == movieDetail.title);
     if (!exists) {
       final docRef = await _firestore.collection('wishlist').add({
+        'userId': userId,
         'title': movieDetail.title,
         'posterPath': movieDetail.posterPath,
         'voteAverage': movieDetail.voteAverage,
         'genreIds': movieDetail.genres.map((g) => g.id.toString()).join(','),
         'runtime': movieDetail.runtime,
       });
-      state = [...state, WishlistMovie.fromMap(docRef.id, movieDetail.toMap())];
+
+      final newMovie = WishlistMovie(
+        id: docRef.id,
+        userId: userId!,
+        title: movieDetail.title,
+        posterPath: movieDetail.posterPath,
+        voteAverage: movieDetail.voteAverage,
+        genreIds: movieDetail.genres.map((g) => g.id).toList(),
+        runtime: movieDetail.runtime,
+      );
+
+      state = [...state, newMovie];
     }
   }
 
   Future<void> removeFromWishlist(String title) async {
-    final movie = state.firstWhere((m) => m.title == title,
-        orElse: () => WishlistMovie(
-            id: '', title: '', posterPath: '', voteAverage: 0.0, genreIds: []));
+    if (userId == null) return;
+
+    final movie = state.firstWhere(
+      (m) => m.title == title,
+      orElse: () => WishlistMovie(
+        id: '',
+        userId: '',
+        title: '',
+        posterPath: '',
+        voteAverage: 0.0,
+        genreIds: [],
+      ),
+    );
+
     if (movie.id.isNotEmpty) {
       await _firestore.collection('wishlist').doc(movie.id).delete();
       state = state.where((m) => m.id != movie.id).toList();
